always true
piles.length is even, which means Alex pick half and Lee pick half. So we have many different divisions here.
But to make it simple, we find the fact that sum(piles[even]) and sum(piles[odd]) must not be equal.
That's why we want to pick all even indexes or all odd indexes to get the larger half.
And Alex always have the right to choose all even indexes or all odd indexes because he starts first.
Then Alex can always to get the larger half.